refactor(qseow): Split up too large source files to improve maintainability and testability
